net: lib: download_client: Add support for TCP receive timeout
It should be possible to set receive timeout also for
TCP sockets. For example if device goes out of LTE
coverage during the download, the recv() call will
block for a very long time if receive timeout has not
been set. In case of a receive timeout with TCP,
the request is not retransmitted.

Setting of the socket timeout is moved to
client_connect(), because earlier the timeout option
was incorrectly not set during reconnect. Also the
decision to set the socket receive timeout is done
based on the used protocol, not based on CONFIG_COAP
(even though COAP has been enabled, the download may
still use HTTP).

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>
Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>