drivers: lcz_fem: Add driver for CE, test, and errata.
The CE mode changes the gain of the FEM using a power model.
It requires the FEM to be used in SPI mode.

Add errata model to handle production error that occurred in Feb 2023.
Add test power tables to simplify obtained 20 dBm baseline.

Allow region to be set using Kconfig and board file.

L2-197