drivers: lcz_fem: Update driver for CE and errata
The CE mode changes the gain of the FEM using a power model.
It requires the FEM to be used in SPI mode.
Modify Kconfig to select Laird Connectivity FEM power model.
L2-208

Add errata model to handle production error that occurred in Feb 2023.
It also requires the FEM to be used in SPI mode.
Add test power tables to simplify obtaining 20 dBm baseline.

Allow region to be set using Kconfig and board file.
L2-197

Signed-off-by: Andrew Hedin <andrew.hedin@lairdconnect.com>