drivers: lcz_fem: Add Laird Connectivity FEM driver
This driver is used by the BL5340PA module that contains
a nRF5340 and nRF21540.

The Laird Connectivity driver works in conjunction with the
Nordic driver. It reduces the RF power output to be compliant
with regulatory requirements. It controls the SPI chip select and
antenna select (of the FEM) in accordance
with the recommendations of the nRF21540 errata.

Signed-off-by: Andrew Hedin <andrew.hedin@ezurio.com>